prerender and check_pet_tp addition
Added register_event-prerender and check_pet_tp() function for the purpose of automatically changing to the Automaton WSD on Master whenever the automaton has over 1000TP.  Also added tp_threshold_reached object to check if the add_to_chat message was sent only once. Without that check it will spam in your ingame chat. This is a big lua script update to Puppetmaster. This is like...The missing logic requested on the bg-wiki Puppetmaster Guide.